+++ ChangeLog.rst
@@ -6,6 +6,8 @@ ChangeLog
 
 - Add support for Python 3.6
 - Drop obsolete .grocker file
+- Switch to docker(-py) 2
+- Only manage Grocker created stuff when purging
 
 4.6 (2016-12-22)
 ----------------